adaptive_ai_nutrition/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openai>=1.14.0
elasticsearch==8.13.0
python-telegram-bot==21.1.1
pytest==8.1.1
httpx==0.27.0
